농산물 분류별 data update 수정

+++ update_data.py
@@ -5,15 +5,15 @@ from datetime import datetime, timedelta
 
 import load_data
 
-
 def update_weather_csv():
     station_info = pd.read_csv('./input/종관기상_관측지점.csv')
 
     today = datetime.now()
     yesterday = today - timedelta(days=1)
+    current_year = yesterday.year
+
     yesterday = yesterday.strftime("%Y%m%d")
 
-    current_year = datetime.now().year
     for idx, row in station_info.iterrows():
         stn_id = row['지점코드']
         filename = f'./output/weather/{stn_id}_{current_year}.csv'
@@ -23,31 +23,42 @@ def update_weather_csv():
             update_weather = pd.concat([exists_weather, update_weather], ignore_index=True)
             update_weather = update_weather.drop_duplicates()
             update_weather.to_csv(filename, index=False)
+        else:
+            update_weather = load_data.request_weather_api(stn_id, yesterday)
+            update_weather.to_csv(filename, index=False)
 
 
-def update_price_csv():
-    filename = './output/price_data.csv'
-
+def update_price_csv(code, value):
     today = datetime.now()
+    current_year = today.year
+
     today = today.strftime("%Y%m%d")
 
-    current_data = load_data.request_price_api(today)
+    filename = f"./output/price/{value}_{current_year}.csv"
 
-    past_data = pd.read_csv(filename)
-
-    update_data = pd.concat([past_data, current_data], ignore_index=True)
-    update_data.to_csv(filename, index=False)
+    current_data = load_data.request_price_api(today, code)
+    if os.path.exists(filename):
+        past_data = pd.read_csv(filename)
+        update_data = pd.concat([past_data, current_data], ignore_index=True)
+        update_data = update_data.drop_duplicates()
+        update_data.to_csv(filename, index=False)
+    else:
+        current_data.to_csv(filename, index=False)
 
     return current_data
 
-def update_price_issue(current_data):
+def update_price_issue(current_data, value):
     today = datetime.now()
 
-    title = f'{today.strftime("%Y%m%d")} - 서울 채소류 소매 가격 정보'
+    title = f'전국 소매 가격 정보 - {today.strftime("%Y%m%d")}'
     GITHUB_TOKEN = os.environ['MY_GITHUB_TOKEN']
     REPO_NAME = 'Action_API'
     repo = Github(GITHUB_TOKEN).get_user().get_repo(REPO_NAME)
     body = current_data
+
+    def add_comment_to_issue(repo, issue_number, comment):
+        issue = repo.get_issue(issue_number)
+        issue.create_comment(comment)
 
     if REPO_NAME == repo.name:
         if not body.empty:
@@ -57,16 +68,28 @@ def update_price_issue(current_data):
             for index, row in body.iterrows():
                 table_md += '| ' + ' | '.join([str(cell) for cell in row]) + ' |\n'
 
-            body = table_md
+            body = f"### {value}\n" + table_md
         else:
             body = '데이터가 없습니다.'
 
-        repo.create_issue(title=title, body=body)
+        existing_issue = None
+        for issue in repo.get_issues():
+            if issue.title == title:
+                existing_issue = issue
+                break
+
+        if existing_issue:
+            add_comment_to_issue(repo, existing_issue.number, body)
+        else:
+            repo.create_issue(title=title, body=body)
 
 def main():
-    current_data = update_price_csv()
+    code_dict = {100: '식량작물', 200: '채소류', 300: '특용작물', 400: '과일류', 500: '축산물', 600: '수산물'}
+    for code, value in code_dict.items():
+        current_data = update_price_csv(code, value)
+        update_price_issue(current_data, value)
+
     update_weather_csv()
-    update_price_issue(current_data)
 
 if __name__ == '__main__':
     main()
